Make IResource responses from controller serializable
When using a controller that returns a resource interface instead of a
representation implementation the object is not serialized properly into
hal+json. Specificly the list is a collection of link objects instead of
a list object with link properties. Changing the type that IsResource
and CanWriteType use from Representation to IResource allows for the
desired behavior.