Added a class AttributeRepresentation which inherits from Representation and uses some helper functions to set the values for Rel and Href based on custom attributes added to the Representation class.  It also builds the list of links using custom attributes.
Also modified the ResourceConverter class so that it can use the Rel set by a parent object rather than a default Rel value for embedded resources.  This allows a parent resource to set the Rel value of a sub-resource or collection of resources so the resource name can be controlled by the parent.  This helps in scenarios where you may have an Org class with a Janitor and a Boss property.  If both Janitor and Boss are of type employee, then they would both be serialized with the "employee"  name.  Now, the Org and specify that the Rel for Janitor is "janitor" and the Rel for Boss is "boss" so they get serialized with those names and can be distinguished easier by the client consuming the response.